Rename allocator/metrics to allocator/balanced
