+++ allocator/balanced/balanced_test.go
@@ -1,4 +1,4 @@
-package metrics
+package balanced
 
 import (
 	"context"
