+++ allocator/balanced/config.go
@@ -1,4 +1,4 @@
-package metrics
+package balanced
 
 import (
 	"encoding/json"
