+++ allocator/balanced/config_test.go
@@ -1,4 +1,4 @@
-package metrics
+package balanced
 
 import (
 	"os"
