+++ cmdutils/configs.go
@@ -8,7 +8,7 @@ import (
 	"github.com/pkg/errors"
 
 	ipfscluster "github.com/ipfs/ipfs-cluster"
-	"github.com/ipfs/ipfs-cluster/allocator/metrics"
+	"github.com/ipfs/ipfs-cluster/allocator/balanced"
 	"github.com/ipfs/ipfs-cluster/api/ipfsproxy"
 	"github.com/ipfs/ipfs-cluster/api/rest"
 	"github.com/ipfs/ipfs-cluster/config"
@@ -35,7 +35,7 @@ type Configs struct {
 	Crdt             *crdt.Config
 	Statelesstracker *stateless.Config
 	Pubsubmon        *pubsubmon.Config
-	MetricsAlloc     *metrics.Config
+	BalancedAlloc    *balanced.Config
 	Diskinf          *disk.Config
 	Numpininf        *numpin.Config
 	Tagsinf          *tags.Config
@@ -224,7 +224,7 @@ func (ch *ConfigHelper) init() {
 		Crdt:             &crdt.Config{},
 		Statelesstracker: &stateless.Config{},
 		Pubsubmon:        &pubsubmon.Config{},
-		MetricsAlloc:     &metrics.Config{},
+		BalancedAlloc:    &balanced.Config{},
 		Diskinf:          &disk.Config{},
 		Numpininf:        &numpin.Config{},
 		Tagsinf:          &tags.Config{},
@@ -239,7 +239,7 @@ func (ch *ConfigHelper) init() {
 	man.RegisterComponent(config.IPFSConn, cfgs.Ipfshttp)
 	man.RegisterComponent(config.PinTracker, cfgs.Statelesstracker)
 	man.RegisterComponent(config.Monitor, cfgs.Pubsubmon)
-	man.RegisterComponent(config.Allocator, cfgs.MetricsAlloc)
+	man.RegisterComponent(config.Allocator, cfgs.BalancedAlloc)
 	man.RegisterComponent(config.Informer, cfgs.Diskinf)
 	// man.RegisterComponent(config.Informer, cfgs.Numpininf)
 	man.RegisterComponent(config.Informer, cfgs.Tagsinf)
